- Updated to generate enum ScanForHardwareChanges - Replaced PostBuildLibExport the BuildEventer tool - Removed the obsolete vtStorProtocolManaged
