+++ vtStor.Test/vtStor.Test.vcxproj
@@ -75,22 +75,18 @@
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
-    <LinkIncremental>true</LinkIncremental>
     <OutDir>$(ProjectDir)$(Platform)$(Configuration)\</OutDir>
     <IntDir>$(ProjectDir)$(Platform)$(Configuration)\Intermediate\</IntDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <LinkIncremental>true</LinkIncremental>
     <OutDir>$(ProjectDir)$(Platform)$(Configuration)\</OutDir>
     <IntDir>$(ProjectDir)$(Platform)$(Configuration)\Intermediate\</IntDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
-    <LinkIncremental>true</LinkIncremental>
     <OutDir>$(ProjectDir)$(Platform)$(Configuration)\</OutDir>
     <IntDir>$(ProjectDir)$(Platform)$(Configuration)\Intermediate\</IntDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <LinkIncremental>true</LinkIncremental>
     <OutDir>$(ProjectDir)$(Platform)$(Configuration)\</OutDir>
     <IntDir>$(ProjectDir)$(Platform)$(Configuration)\Intermediate\</IntDir>
   </PropertyGroup>
