+++ vtStorAta/Common.props
@@ -16,7 +16,8 @@
       <AdditionalLibraryDirectories>$(SolutionDir)StorageUtility\$(Platform)$(Configuration);$(SolutionDir)DriveAta\$(Platform)$(Configuration);$(SolutionDir)vtStor\$(Platform)$(Configuration);%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
     </Link>
     <PostBuildEvent>
-      <Command>python $(SolutionDir)PostBuildLibExport.py $(ProjectName) $(Platform)$(Configuration) $(SolutionDir)</Command>
+      <Command>cd $(SolutionDir)
+python Build\BuildEventer.py Build\BuildEventer.xml $(ProjectName) $(Platform) $(Configuration)</Command>
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup />
